Added nanodiscAnalysis.tcl
This is the first version of the analysis tools. These are a work in progress. Contains functions to measure nanodisc diameter so far. Will continue to add to these procedures. Requires procedures.tcl for cartToPolar2D function. Can consider adding this to the main body of the analysis script at a later date.